project/requirements.txt
pillow
numpy
scikit-learn
plotly
streamlit
pandas
